requirements.txt
Flask
Flask-SQLAlchemy
flask-serialize
psycopg2-binary
Pillow
requests
